feat(plugins): moved base constraints classes to keel-api
