+++ keel-api/src/main/kotlin/com/netflix/spinnaker/keel/api/Constraint.kt
@@ -1,0 +1,4 @@
+package com.netflix.spinnaker.keel.api
+
+abstract class Constraint(val type: String)
+abstract class StatefulConstraint(type: String) : Constraint(type)
