+++ keel-core/src/main/kotlin/com/netflix/spinnaker/keel/api/DependsOnConstraint.kt
@@ -1,0 +1,9 @@
+package com.netflix.spinnaker.keel.api
+
+/**
+ * A constraint that requires that an artifact has been successfully promoted to a previous
+ * environment first.
+ */
+data class DependsOnConstraint(
+  val environment: String
+) : Constraint("depends-on")
