+++ keel-core/src/main/kotlin/com/netflix/spinnaker/keel/api/ManualJudgementConstraint.kt
@@ -1,0 +1,7 @@
+package com.netflix.spinnaker.keel.api
+
+import java.time.Duration
+
+data class ManualJudgementConstraint(
+  val timeout: Duration = Duration.ofDays(7)
+) : StatefulConstraint("manual-judgement")
