+++ keel-core/src/test/kotlin/com/netflix/spinnaker/keel/constraints/DependsOnConstraintEvaluatorTests.kt
@@ -25,7 +25,7 @@ internal class DependsOnConstraintEvaluatorTests : JUnit5Minutests {
     val constrainedEnvironment = Environment(
       name = "staging",
       constraints = setOf(
-        DependsOnConstraint("test")
+        DependsOnConstraint(environment = "test")
       )
     )
     val previousEnvironment = Environment(
